tasks-qute: Rename QuarkusTest classes to *Test.java
